Split engine capi into another crate
