+++ engine/capi/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "kime-engine-capi"
+version = "0.1.0"
+authors = ["Riey <creeper844@gmail.com>"]
+edition = "2018"
+license = "GPL-3.0"
+
+[lib]
+name = "kime_engine"
+crate-type = ["cdylib"]
+
+[dependencies]
+kime-engine-core = { path = "../core" }
+
