+++ engine/core/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "kime-engine"
+name = "kime-engine-core"
 version = "0.4.1"
 authors = ["Riey <creeper844@gmail.com>"]
 edition = "2018"
 license = "GPL-3.0"
-
-[lib]
-crate-type = ["cdylib"]
 
 [dependencies]
 ahash = { version = "0.6.2", features = ["serde"] }
