+++ engine/core/src/characters.rs
