+++ engine/core/src/keycode.rs
@@ -172,12 +172,10 @@ impl Key {
         Self::new(code, false, false, false)
     }
 
-    #[cfg(test)]
     pub const fn shift(code: KeyCode) -> Self {
         Self::new(code, true, false, false)
     }
 
-    #[cfg(test)]
     pub const fn ctrl(code: KeyCode) -> Self {
         Self::new(code, false, true, false)
     }
