--- engine/src/tests.rs
@@ -1,3 +1,0 @@
-mod dubeolsik;
-mod sebeolsik_390;
-mod sebeolsik_391;
