+++ gtk3/Cargo.toml
@@ -13,5 +13,5 @@ gdk-sys = "0.10.0"
 glib-sys = "0.10.1"
 gobject-sys = "0.10.0"
 gtk-sys = "0.10.0"
-kime-engine-cffi = { path = "../engine-cffi" }
+kime-engine-cffi = { path = "../engine/cffi" }
 pango-sys = "0.10.0"
